Reimplement by DataBindingAdapter that TextView's 'Web Search' menu function.
